Refactor old contact pages, move index code to parent contact page except form. Needs testing and refinement
