+++ app/routes/_front.contact._index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { useLoaderData, useActionData, Link } from "@remix-run/react";
+import { Form, useActionData, Link } from "@remix-run/react";
 import {
   json,
   redirect,
@@ -10,7 +10,6 @@ import {
 
 import { BASE_API_URL } from "~/common/constants.server";
 import { getGlobalEnv } from "~/common/utils";
-import { SlidingHeaderPage } from "~/components/pages";
 import { FaCaretDown, FaArrowDown, FaArrowRightLong } from "react-icons/fa6";
 import { FaMobileAlt } from "react-icons/fa";
 import { HeadingOne } from "~/components/headings";
@@ -78,7 +77,7 @@ const contactMenu = [
 
 function validateRequired(value: unknown): string[] {
   if (!value) {
-    return [`This field is required`];
+    return [MESSAGES["ja"].form.required];
   }
   return [];
 }
@@ -160,10 +159,11 @@ export const action = async ({ request }: ActionArgs) => {
         { status: 400 }
       );
     }
+    console.log("Got here");
     return redirect("/contact/success");
   } catch (error) {
     // network error
-    errors.non_field_errors = [MESSAGES["en"].form.standard400];
+    errors.non_field_errors = [MESSAGES["ja"].form.standard400];
     return json(
       {
         fields,
@@ -175,25 +175,7 @@ export const action = async ({ request }: ActionArgs) => {
   } // catch
 }; // action
 
-export async function loader() {
-  try {
-    const apiUrl = `${BASE_API_URL}/pages/?slug=contact&type=contacts.ContactPage&fields=*`;
-    const response = await fetch(apiUrl);
-    const contactPageData = await response.json();
-    if (!response.ok || !contactPageData.items.length) {
-      throw new Response("Sorry, that is a 404", { status: 404 });
-    }
-    const page = contactPageData.items[0];
-    return json({ page });
-  } catch (error) {
-    throw new Response("Oh sorry, that is a 500", { status: 500 });
-  }
-} // loader
-
 export default function ContactIndexPage() {
-  const ENV = getGlobalEnv();
-  const { page } = useLoaderData<typeof loader>();
-
   const actionData = useActionData<typeof action>();
   const maxMsgLength = 300;
   const [remaining, setRemaining] = React.useState(maxMsgLength);
@@ -213,404 +195,183 @@ export default function ContactIndexPage() {
   }, []);
 
   return (
-    <SlidingHeaderPage
-      mainTitle={page.title}
-      subTitle={page.display_title}
-      swooshBackColor="cream"
-      swooshFrontColor="beige"
+    <Form
+      preventScrollReset
+      action="/contact?index"
+      className="ct-form__form"
+      noValidate
+      method="post"
     >
-      <div className="ct-nav">
-        {contactMenu.map((item) => {
-          return <PageNav key={item.id} text={item.text} url={item.url} />;
-        })}
+      {actionData?.errors?.non_field_errors?.length ? (
+        <div className="g-form__nonfield-errors">
+          <ul>
+            {actionData.errors.non_field_errors.map((error) => (
+              <li role="alert" key={error}>
+                {error}
+              </li>
+            ))}
+          </ul>
+        </div>
+      ) : null}
+      <div className="g-form__input-group">
+        <label
+          className="g-form__text-label g-required"
+          htmlFor="full-name-input"
+        >
+          お名前（日本語表記）
+        </label>
+        <input
+          type="text"
+          id="full-name-input"
+          name="fullName"
+          required
+          placeholder="例: 山田太郎 | ex. Bob Jones"
+          defaultValue={actionData?.fields?.fullName}
+          aria-invalid={Boolean(actionData?.errors?.fullName?.length)}
+          aria-errormessage={
+            actionData?.errors?.fullName?.length
+              ? "full-name-errors"
+              : undefined
+          }
+        />
+        {actionData?.errors?.fullName?.length ? (
+          <ul
+            className="g-form__validation-errors"
+            role="alert"
+            id="full-name-errors"
+          >
+            {actionData.errors.fullName.map((error: string) => {
+              return <li key={error}>{error}</li>;
+            })}
+          </ul>
+        ) : null}
       </div>
 
-      <section id="trial">
-        <div className="ct-texp">
-          <div className="g-narrow-container">
-            <div className="ct-texp__heading">
-              <HeadingOne
-                enText={page.trial_en_title}
-                jpText={page.trial_jp_title}
-                align="center"
-                bkground="light"
-                level="h2"
-              />
-            </div>
-            <div
-              className="ct-texp__intro"
-              dangerouslySetInnerHTML={{ __html: page.trial_intro }}
-            />
-            <div className="ct-texp__steps">
-              {page.trial_steps.map((step, i, arr) => {
-                const len = arr.length;
-                return (
-                  <div key={step.id}>
-                    <NumberedHorizontalCards
-                      number={`0${i + 1}`}
-                      jaTitle={step.value.title}
-                      text={step.value.text}
-                      src={
-                        step.value.image
-                          ? `${ENV.BASE_BACK_URL}${step.value.image.thumbnail.src}`
-                          : null
-                      }
-                      alt={
-                        step.value.image ? step.value.image.medium.alt : null
-                      }
-                    />
-                    {i + 1 < len ? (
-                      <div className="ct-texp__step__caret">
-                        <FaCaretDown />
-                      </div>
-                    ) : null}
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="experience">
-        <div className="ct-texp">
-          <div className="g-narrow-container">
-            <div className="ct-texp__heading">
-              <HeadingOne
-                enText={page.exp_en_title}
-                jpText={page.exp_jp_title}
-                align="center"
-                bkground="light"
-                level="h2"
-              />
-            </div>
-            <div
-              className="ct-texp__intro"
-              dangerouslySetInnerHTML={{ __html: page.trial_intro }}
-            />
-            <div className="ct-texp__steps">
-              {page.trial_steps.map((step, i, arr) => {
-                const len = arr.length;
-                return (
-                  <div key={step.id}>
-                    <NumberedHorizontalCards
-                      number={`0${i + 1}`}
-                      jaTitle={step.value.title}
-                      text={step.value.text}
-                      src={
-                        step.value.image
-                          ? `${ENV.BASE_BACK_URL}${step.value.image.thumbnail.src}`
-                          : null
-                      }
-                      alt={
-                        step.value.image ? step.value.image.medium.alt : null
-                      }
-                    />
-                    {i + 1 < len ? (
-                      <div className="ct-texp__step__caret">
-                        <FaCaretDown />
-                      </div>
-                    ) : null}
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="q&a">
-        <div className="ct-qas">
-          <div className="g-narrow-container">
-            <div className="ct-qa__heading">
-              <HeadingOne
-                enText={page.qa_en_title}
-                jpText={page.qa_jp_title}
-                align="center"
-                bkground="light"
-                level="h2"
-              />
-            </div>
-
-            <div className="ct-qa__qas">
-              {page.qas.map((block) => {
-                return (
-                  <div className="ct-qas__qa" key={block.id}>
-                    <p>{block.value.question}</p>
-                    <p>{block.value.answer}</p>
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="telephone">
-        <div className="ct-tel">
-          <div className="g-narrow-container">
-            <div className="ct-tel__heading">
-              <HeadingOne
-                enText={page.tel_en_title}
-                jpText={page.tel_jp_title}
-                align="center"
-                bkground="light"
-                level="h2"
-              />
-            </div>
-
-            <div className="ct-tel__details">
-              <Link className="ct-tel__link" to={`tel:${page.tel_number}`}>
-                <FaMobileAlt />
-                {page.tel_display_number}
-              </Link>
-              <div className="ct-tel__times">
-                <div>
-                  <span>営業時間</span>
-                </div>
-                <div>
-                  <span>火/木/金</span>
-                  <span>13:00-17:00</span>
-                </div>
-                <div>
-                  <span>水</span>
-                  <span>09:00-17:00</span>
-                </div>
-                <div>
-                  <span>土</span>
-                  <span>09:00-17:00</span>
-                </div>
-                <div>
-                  <span>日/月/祝</span>
-                  <span>休業</span>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <section id="form">
-        <div className="ct-form">
-          <div className="g-narrow-container">
-            <div className="ct-form__heading">
-              <HeadingOne
-                enText={page.form_en_title}
-                jpText={page.form_jp_title}
-                align="center"
-                bkground="light"
-                level="h2"
-              />
-            </div>
-            <div className="ct-form__wrapper">
-              <div className="ct-form__msg">
-                <Link to="#trial">体験レッスンの流れ</Link>や
-                <Link to="#experience">ラーニング・エクスペリエンスの流れ</Link>
-                についてもっと知りたいですか？お問い合わせは火曜日〜土曜日に受け付けています。お問い合わせにはできるだけ早く対応させていただきます。
-              </div>
-              <form
-                action="/contact?index"
-                className="ct-form__form"
-                noValidate
-                method="post"
-              >
-                <div className="g-form__input-group">
-                  <label
-                    className="g-form__text-label g-required"
-                    htmlFor="full-name-input"
-                  >
-                    お名前（日本語表記）
-                  </label>
-                  <input
-                    type="text"
-                    id="full-name-input"
-                    name="fullName"
-                    required
-                    placeholder="例: 山田太郎 | ex. Bob Jones"
-                    defaultValue={actionData?.fields?.fullName}
-                    aria-invalid={Boolean(actionData?.errors?.fullName?.length)}
-                    aria-errormessage={
-                      actionData?.errors?.fullName?.length
-                        ? "full-name-errors"
-                        : undefined
-                    }
-                  />
-                  {actionData?.errors?.fullName?.length ? (
-                    <ul
-                      className="g-form__validation-errors"
-                      role="alert"
-                      id="full-name-errors"
-                    >
-                      {actionData.errors.fullName.map((error: string) => {
-                        return <li key={error}>{error}</li>;
-                      })}
-                    </ul>
-                  ) : null}
-                </div>
-
-                <div className="g-form__input-group">
-                  <label
-                    className="g-form__text-label g-required"
-                    htmlFor="full-en-name-input"
-                  >
-                    お名前（英語表記）
-                  </label>
-                  <input
-                    type="text"
-                    id="full-en-name-input"
-                    name="fullEnName"
-                    required
-                    placeholder="例: Yamada Taro | ex. Bob Jones"
-                    defaultValue={actionData?.fields?.fullEnName}
-                    aria-invalid={Boolean(
-                      actionData?.errors?.fullEnName?.length
-                    )}
-                    aria-errormessage={
-                      actionData?.errors?.fullEnName?.length
-                        ? "full-en-name-errors"
-                        : undefined
-                    }
-                  />
-                  {actionData?.errors?.fullEnName?.length ? (
-                    <ul
-                      className="g-form__validation-errors"
-                      role="alert"
-                      id="full-en-name-errors"
-                    >
-                      {actionData.errors.fullEnName.map((error: string) => {
-                        return <li key={error}>{error}</li>;
-                      })}
-                    </ul>
-                  ) : null}
-                </div>
-
-                <div className="g-form__input-group">
-                  <label
-                    className="g-form__text-label g-required"
-                    htmlFor="email-input"
-                  >
-                    Eメールアドレス
-                  </label>
-                  <input
-                    type="email"
-                    id="email-input"
-                    name="email"
-                    placeholder="例: y.t@abcdz.com | ex. j.b@abcdz.com"
-                    required
-                    defaultValue={actionData?.fields?.email}
-                    aria-invalid={Boolean(actionData?.errors?.email?.length)}
-                    aria-errormessage={
-                      actionData?.errors?.email?.length
-                        ? "email-errors"
-                        : undefined
-                    }
-                  />
-                  {actionData?.errors?.email?.length ? (
-                    <ul
-                      className="g-form__validation-errors"
-                      role="alert"
-                      id="email-errors"
-                    >
-                      {actionData.errors.email.map((error: string) => {
-                        return <li key={error}>{error}</li>;
-                      })}
-                    </ul>
-                  ) : null}
-                </div>
-
-                <div className="g-form__input-group ct-form__textarea">
-                  <label
-                    className="g-form__text-label g-required"
-                    htmlFor="message-input"
-                  >
-                    お問い合わせ内容
-                    <span
-                      className={`ct-form__textarea__remaining ${
-                        remaining > 0 ? "yes" : "no"
-                      }`}
-                    >
-                      {remaining}/300 文字
-                    </span>
-                  </label>
-                  <textarea
-                    id="message-input"
-                    maxLength={maxMsgLength}
-                    name="message"
-                    onChange={messageChangeHandler}
-                    ref={messageRef}
-                    defaultValue={actionData?.fields?.message}
-                    aria-invalid={Boolean(actionData?.errors?.message?.length)}
-                    aria-errormessage={
-                      actionData?.errors?.message?.length
-                        ? "message-errors"
-                        : undefined
-                    }
-                  ></textarea>
-                  {actionData?.errors?.message?.length ? (
-                    <ul
-                      className="form-validation-errors"
-                      role="alert"
-                      id="message-errors"
-                    >
-                      {actionData.errors.message.map((error: string) => {
-                        return <li key={error}>{error}</li>;
-                      })}
-                    </ul>
-                  ) : null}
-                </div>
-
-                <div className="ct-form__privacy">
-                  <div>
-                    <input
-                      type="checkbox"
-                      id="privacy-read"
-                      name="privacy-read"
-                      checked={privacyRead}
-                      autoComplete="off"
-                      onChange={() => setPrivacyRead((prev) => !prev)}
-                    />
-                    <label htmlFor="privacy-read">
-                      プライバシーポリシーに同意する
-                    </label>
-                  </div>
-                  <div>
-                    当社のプライバシーポリシーは
-                    <Link to="/privacy-policy">こちら</Link>
-                  </div>
-                </div>
-                <div className="ct-form__submit">
-                  <button disabled={!privacyRead} type="submit">
-                    ログイン
-                    <FaArrowRightLong />
-                  </button>
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-      </section>
-    </SlidingHeaderPage>
-  );
-}
-
-/*
- * Components
- */
-
-type TPageNavProps = {
-  text: string;
-  url: string;
-};
-
-function PageNav({ text, url }: TPageNavProps) {
-  return (
-    <Link className="ct-nav__link " to={url}>
-      <div className="ct-nav__btn">
-        <div className="ct-nav__icon">
-          <FaArrowDown />
-        </div>
-        <div className="ct-nav__text">{text}</div>
+      <div className="g-form__input-group">
+        <label
+          className="g-form__text-label g-required"
+          htmlFor="full-en-name-input"
+        >
+          お名前（英語表記）
+        </label>
+        <input
+          type="text"
+          id="full-en-name-input"
+          name="fullEnName"
+          required
+          placeholder="例: Yamada Taro | ex. Bob Jones"
+          defaultValue={actionData?.fields?.fullEnName}
+          aria-invalid={Boolean(actionData?.errors?.fullEnName?.length)}
+          aria-errormessage={
+            actionData?.errors?.fullEnName?.length
+              ? "full-en-name-errors"
+              : undefined
+          }
+        />
+        {actionData?.errors?.fullEnName?.length ? (
+          <ul
+            className="g-form__validation-errors"
+            role="alert"
+            id="full-en-name-errors"
+          >
+            {actionData.errors.fullEnName.map((error: string) => {
+              return <li key={error}>{error}</li>;
+            })}
+          </ul>
+        ) : null}
       </div>
-    </Link>
+
+      <div className="g-form__input-group">
+        <label className="g-form__text-label g-required" htmlFor="email-input">
+          Eメールアドレス
+        </label>
+        <input
+          type="email"
+          id="email-input"
+          name="email"
+          placeholder="例: y.t@abcdz.com | ex. j.b@abcdz.com"
+          required
+          defaultValue={actionData?.fields?.email}
+          aria-invalid={Boolean(actionData?.errors?.email?.length)}
+          aria-errormessage={
+            actionData?.errors?.email?.length ? "email-errors" : undefined
+          }
+        />
+        {actionData?.errors?.email?.length ? (
+          <ul
+            className="g-form__validation-errors"
+            role="alert"
+            id="email-errors"
+          >
+            {actionData.errors.email.map((error: string) => {
+              return <li key={error}>{error}</li>;
+            })}
+          </ul>
+        ) : null}
+      </div>
+
+      <div className="g-form__input-group ct-form__textarea">
+        <label
+          className="g-form__text-label g-required"
+          htmlFor="message-input"
+        >
+          お問い合わせ内容
+          <span
+            className={`ct-form__textarea__remaining ${
+              remaining > 0 ? "yes" : "no"
+            }`}
+          >
+            {remaining}/300 文字
+          </span>
+        </label>
+        <textarea
+          id="message-input"
+          maxLength={maxMsgLength}
+          name="message"
+          onChange={messageChangeHandler}
+          ref={messageRef}
+          defaultValue={actionData?.fields?.message}
+          aria-invalid={Boolean(actionData?.errors?.message?.length)}
+          aria-errormessage={
+            actionData?.errors?.message?.length ? "message-errors" : undefined
+          }
+        ></textarea>
+        {actionData?.errors?.message?.length ? (
+          <ul
+            className="g-form__validation-errors"
+            role="alert"
+            id="message-errors"
+          >
+            {actionData.errors.message.map((error: string) => {
+              return <li key={error}>{error}</li>;
+            })}
+          </ul>
+        ) : null}
+      </div>
+
+      <div className="ct-form__privacy">
+        <div>
+          <input
+            type="checkbox"
+            id="privacy-read"
+            name="privacy-read"
+            checked={privacyRead}
+            autoComplete="off"
+            onChange={() => setPrivacyRead((prev) => !prev)}
+          />
+          <label htmlFor="privacy-read">プライバシーポリシーに同意する</label>
+        </div>
+        <div>
+          当社のプライバシーポリシーは
+          <Link to="/privacy-policy">こちら</Link>
+        </div>
+      </div>
+      <div className="ct-form__submit">
+        <button disabled={!privacyRead} type="submit">
+          ログイン
+          <FaArrowRightLong />
+        </button>
+      </div>
+    </Form>
   );
 }
