+++ app/routes/_front.contact.tsx
@@ -1,6 +1,5 @@
 import type { LinksFunction } from "@remix-run/node";
-import { useLoaderData, Outlet } from "@remix-run/react";
-import React from "react";
+import { useLoaderData, Outlet, Link } from "@remix-run/react";
 import { json } from "@remix-run/node";
 import { getGlobalEnv } from "~/common/utils";
 
@@ -8,8 +7,19 @@ import contactStyles from "~/styles/contact.css";
 import { BASE_API_URL } from "~/common/constants.server";
 import pageCStyles from "~/styles/components/pages.css";
 import cardStyles from "~/styles/components/cards.css";
+import { SlidingHeaderPage } from "~/components/pages";
+import { FaCaretDown, FaArrowDown } from "react-icons/fa6";
+import { FaMobileAlt } from "react-icons/fa";
+import { HeadingOne } from "~/components/headings";
+import { NumberedHorizontalCards } from "~/components/cards";
 
-/* Types */
+/**
+ * Types
+ */
+
+/**
+ *  Utils and helper functions
+ */
 
 export const links: LinksFunction = () => [
   { rel: "stylesheet", href: contactStyles },
@@ -17,10 +27,280 @@ export const links: LinksFunction = () => [
   { rel: "stylesheet", href: cardStyles },
 ];
 
+const contactMenu = [
+  {
+    id: 1,
+    text: "無料体験レッスン",
+    url: "#trial",
+  },
+  {
+    id: 2,
+    text: "ランゲージ・エクスペリエンス",
+    url: "#experience",
+  },
+  {
+    id: 3,
+    text: "よくある質問",
+    url: "#q&a",
+  },
+  {
+    id: 4,
+    text: "お電話でのお問い合わせ",
+    url: "#telephone",
+  },
+  {
+    id: 5,
+    text: "お問い合わせフォーム",
+    url: "#form",
+  },
+];
+
+/**
+ * Server functions
+ */
+
+export async function loader() {
+  try {
+    const apiUrl = `${BASE_API_URL}/pages/?slug=contact&type=contacts.ContactPage&fields=*`;
+    const response = await fetch(apiUrl);
+    const contactPageData = await response.json();
+    if (!response.ok || !contactPageData.items.length) {
+      throw new Response("Sorry, that is a 404", { status: 404 });
+    }
+    const page = contactPageData.items[0];
+    return json({ page });
+  } catch (error) {
+    throw new Response("Oh sorry, that is a 500", { status: 500 });
+  }
+} // loader
+
 export default function ContactParentPage() {
+  const ENV = getGlobalEnv();
+  const { page } = useLoaderData<typeof loader>();
+
   return (
-    <>
-      <Outlet />
-    </>
+    <SlidingHeaderPage
+      mainTitle={page.title}
+      subTitle={page.display_title}
+      swooshBackColor="cream"
+      swooshFrontColor="beige"
+    >
+      <div className="ct-nav">
+        {contactMenu.map((item) => {
+          return <PageNav key={item.id} text={item.text} url={item.url} />;
+        })}
+      </div>
+
+      <section id="trial">
+        <div className="ct-texp">
+          <div className="g-narrow-container">
+            <div className="ct-texp__heading">
+              <HeadingOne
+                enText={page.trial_en_title}
+                jpText={page.trial_jp_title}
+                align="center"
+                bkground="light"
+                level="h2"
+              />
+            </div>
+            <div
+              className="ct-texp__intro"
+              dangerouslySetInnerHTML={{ __html: page.trial_intro }}
+            />
+            <div className="ct-texp__steps">
+              {page.trial_steps.map((step, i, arr) => {
+                const len = arr.length;
+                return (
+                  <div key={step.id}>
+                    <NumberedHorizontalCards
+                      number={`0${i + 1}`}
+                      jaTitle={step.value.title}
+                      text={step.value.text}
+                      src={
+                        step.value.image
+                          ? `${ENV.BASE_BACK_URL}${step.value.image.thumbnail.src}`
+                          : null
+                      }
+                      alt={
+                        step.value.image ? step.value.image.medium.alt : null
+                      }
+                    />
+                    {i + 1 < len ? (
+                      <div className="ct-texp__step__caret">
+                        <FaCaretDown />
+                      </div>
+                    ) : null}
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+        </div>
+      </section>
+
+      <section id="experience">
+        <div className="ct-texp">
+          <div className="g-narrow-container">
+            <div className="ct-texp__heading">
+              <HeadingOne
+                enText={page.exp_en_title}
+                jpText={page.exp_jp_title}
+                align="center"
+                bkground="light"
+                level="h2"
+              />
+            </div>
+            <div
+              className="ct-texp__intro"
+              dangerouslySetInnerHTML={{ __html: page.trial_intro }}
+            />
+            <div className="ct-texp__steps">
+              {page.trial_steps.map((step, i, arr) => {
+                const len = arr.length;
+                return (
+                  <div key={step.id}>
+                    <NumberedHorizontalCards
+                      number={`0${i + 1}`}
+                      jaTitle={step.value.title}
+                      text={step.value.text}
+                      src={
+                        step.value.image
+                          ? `${ENV.BASE_BACK_URL}${step.value.image.thumbnail.src}`
+                          : null
+                      }
+                      alt={
+                        step.value.image ? step.value.image.medium.alt : null
+                      }
+                    />
+                    {i + 1 < len ? (
+                      <div className="ct-texp__step__caret">
+                        <FaCaretDown />
+                      </div>
+                    ) : null}
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+        </div>
+      </section>
+
+      <section id="q&a">
+        <div className="ct-qas">
+          <div className="g-narrow-container">
+            <div className="ct-qa__heading">
+              <HeadingOne
+                enText={page.qa_en_title}
+                jpText={page.qa_jp_title}
+                align="center"
+                bkground="light"
+                level="h2"
+              />
+            </div>
+
+            <div className="ct-qa__qas">
+              {page.qas.map((block) => {
+                return (
+                  <div className="ct-qas__qa" key={block.id}>
+                    <p>{block.value.question}</p>
+                    <p>{block.value.answer}</p>
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+        </div>
+      </section>
+
+      <section id="telephone">
+        <div className="ct-tel">
+          <div className="g-narrow-container">
+            <div className="ct-tel__heading">
+              <HeadingOne
+                enText={page.tel_en_title}
+                jpText={page.tel_jp_title}
+                align="center"
+                bkground="light"
+                level="h2"
+              />
+            </div>
+
+            <div className="ct-tel__details">
+              <Link className="ct-tel__link" to={`tel:${page.tel_number}`}>
+                <FaMobileAlt />
+                {page.tel_display_number}
+              </Link>
+              <div className="ct-tel__times">
+                <div>
+                  <span>営業時間</span>
+                </div>
+                <div>
+                  <span>火/木/金</span>
+                  <span>13:00-17:00</span>
+                </div>
+                <div>
+                  <span>水</span>
+                  <span>09:00-17:00</span>
+                </div>
+                <div>
+                  <span>土</span>
+                  <span>09:00-17:00</span>
+                </div>
+                <div>
+                  <span>日/月/祝</span>
+                  <span>休業</span>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+
+      <section id="form">
+        <div className="ct-form">
+          <div className="g-narrow-container">
+            <div className="ct-form__heading">
+              <HeadingOne
+                enText={page.form_en_title}
+                jpText={page.form_jp_title}
+                align="center"
+                bkground="light"
+                level="h2"
+              />
+            </div>
+            <div className="ct-form__wrapper">
+              <div className="ct-form__msg">
+                <Link to="#trial">体験レッスンの流れ</Link>や
+                <Link to="#experience">ラーニング・エクスペリエンスの流れ</Link>
+                についてもっと知りたいですか？お問い合わせは火曜日〜土曜日に受け付けています。お問い合わせにはできるだけ早く対応させていただきます。
+              </div>
+              <Outlet />
+            </div>
+          </div>
+        </div>
+      </section>
+    </SlidingHeaderPage>
+  );
+}
+
+/*
+ * Components
+ */
+
+type TPageNavProps = {
+  text: string;
+  url: string;
+};
+
+function PageNav({ text, url }: TPageNavProps) {
+  return (
+    <Link className="ct-nav__link " to={url}>
+      <div className="ct-nav__btn">
+        <div className="ct-nav__icon">
+          <FaArrowDown />
+        </div>
+        <div className="ct-nav__text">{text}</div>
+      </div>
+    </Link>
   );
 }
